// resources/admin/src/lib/apollo/graphql.tsx
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Cursor: { input: any; output: any; }
  DateOnly: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  Time: { input: any; output: any; }
  TimeOnly: { input: any; output: any; }
};
// ...
export type Edge = {
  __typename?: 'Edge';
  belongsToCaption?: Maybe<Scalars['String']['output']>;
  caption: Scalars['String']['output'];
  name: Scalars['String']['output'];
  private?: Maybe<Scalars['Boolean']['output']>;
  relatedEntity: Entity;
  relationType?: Maybe<RelationType>;
  required?: Maybe<Scalars['Boolean']['output']>;
};

export type EdgeWhereUniqueInput = {
  caption?: InputMaybe<Scalars['String']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
};

export type Entity = {
  __typename?: 'Entity';
  caption: Scalars['String']['output'];
  df: Scalars['String']['output'];
  displayField: Field;
  edges?: Maybe<Array<Edge>>;
  fields?: Maybe<Array<Field>>;
  name: Scalars['String']['output'];
  owner: EntityOwner;
  pluralName: Scalars['String']['output'];
};

export type EntityConnectInput = {
  connect: EntityWhereUniqueInput;
};

export enum EntityOwner {
  Admin = 'Admin',
  User = 'User'
}

export type EntityWhereUniqueInput = {
  caption?: InputMaybe<Scalars['String']['input']>;
  name?: InputMaybe<Scalars['String']['input']>;
};

export type Field = {
  __typename?: 'Field';
  acceptedValues?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  caption: Scalars['String']['output'];
  defaultValue?: Maybe<Scalars['String']['output']>;
  max?: Maybe<Scalars['String']['output']>;
  min?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  private?: Maybe<Scalars['Boolean']['output']>;
  required?: Maybe<Scalars['Boolean']['output']>;
  type: Scalars['String']['output'];
  unique?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export enum RelationType {
  Many = 'Many',
  ManyToMany = 'ManyToMany',
  ManyToOne = 'ManyToOne',
  One = 'One',
  OneToMany = 'OneToMany',
  OneToOne = 'OneToOne'
}